requirements.txt
google-cloud-discoveryengine>=0.11.0
google-api-core>=2.11.0
pymupdf4llm>=0.0.5
pymupdf>=1.23.0
requests>=2.31.0
langchain-text-splitters>=0.2.0
pandas>=2.0.0

# Embedding models
sentence-transformers>=2.2.0
numpy>=1.24.0

# Visualization
umap-learn>=0.5.4
plotly>=5.18.0
scikit-learn>=1.3.0
google-genai==1.31.0